requirements.txt
absl-py==0.13.0
adabelief-tf==0.2.1
argcomplete==1.12.3
astunparse==1.6.3
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
cleverhans==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
dm-tree==0.1.6
easydict==1.9
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
importlib-metadata==4.6.4
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==7.0.1
jupyter-core==4.7.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.2
mnist==0.2.2
nest-asyncio==1.5.1
nose==1.3.7
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.2.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-probability==0.13.0
termcolor==1.1.0
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
